// src/app/taf-fiscal/models/item-table-marketing-by-farmer.ts
export interface ItemTableMarketingByFarmer {
  company: string;
  branchId?: string;
  acquiringBranch?: string;
  acquiringCNPJ?: string;
  taxNumber?: string;
  taxNumberFormated: string;
  totalGrossValue: number;
  sociaSecurityContributionValue: number;
  sociaSecurityContributionValueGilrat: number;
  sociaSecurityContributionValueSenar: number;
  key?: string;
  status?: string;
  errors?: string;
  errorId?: string;
  protocol?: string;
}